+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Many bug fixes to coincide with release of Zapit on eLife.
     use only and must be enabled at the command line via `hZP.stimConfig.useClockedDO = true`
     (off by default) once a stimulus config file has been loaded. See [Issue 163](https://github.com/Zapit-Optostim/zapit/issues/163)
   * Subtract background image when running getLaserPosAccuracy. This uses a different threshold.  
+  * Switch camera reads to peekdata, which should be smoother. 
 
 
 Bugfixes